PR FIXUP - Rename description package
